+++ frontend/package.json
@@ -29,6 +29,9 @@
     "typescript": "5.1.6"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "env-cmd": "^10.1.0",
     "jest": "^29.6.2",
     "jest-environment-jsdom": "^29.6.2",
