+++ frontend/src/lib/msw/MSWWrapper.tsx
@@ -1,0 +1,16 @@
+import { ReactNode } from 'react';
+
+export type MSWWrapperProps = {
+  children: ReactNode;
+};
+
+require('./init');
+export const MSWWrapper = ({
+  children,
+}: MSWWrapperProps) => {
+  return (
+    <>
+      { children }
+    </>
+  );
+};
