+++ frontend/src/lib/msw/init/browser.ts
@@ -1,0 +1,4 @@
+import { setupWorker } from 'msw';
+import { handlers } from '../handlers';
+
+export const worker = setupWorker(...handlers);
