+++ frontend/src/pages/_app.tsx
@@ -2,10 +2,10 @@ import { AppProps } from 'next/app';
 import dynamic from 'next/dynamic';
 
 import { API_MOCKING } from '@/config/constants';
-import { MSWWrapperProps } from '@/lib/msw';
+import { MSWWrapperProps } from '@/lib/msw/MSWWrapper';
 
 const MSWWrapper = dynamic<MSWWrapperProps>(() =>
-  import('@/lib/msw').then(({ MSWWrapper }) => MSWWrapper)
+  import('@/lib/msw/MSWWrapper').then(({ MSWWrapper }) => MSWWrapper)
 );
 
 const App = ({ Component, pageProps }: AppProps) => {
